% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{AssetPricingFactor-class}
\alias{AssetPricingFactor-class}
\alias{.AssetPricingFactor}
\title{S4 class for asset pricing factor object}
\description{
S4 class for asset pricing factor object
}
