% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{CHP_factor}
\alias{CHP_factor}
\alias{CHP_factor,FuturesTS,FuturesCFTC-method}
\alias{CHP_factor,AssetPricingFactor}
\title{Commercial hedging pressure (CHP) factor}
\usage{
CHP_factor(price_data, CHP_data, update_frequency = "month",
  return_frequency = "day", ranking_period = 6L,
  long_threshold = 0.5, short_threshold = 0.5, weighted = F)

\S4method{CHP_factor}{FuturesTS,FuturesCFTC}(price_data, CHP_data,
  update_frequency = "month", return_frequency = "day",
  ranking_period = 6L, long_threshold = 0.5, short_threshold = 0.5,
  weighted = F)
}
\arguments{
\item{price_data}{an S4 object of class \code{\linkS4class{FuturesTS}}.
\code{\linkS4class{FuturesTS}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} function in the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{CHP_data}{an S4 object of class \code{\linkS4class{FuturesCFTC}}.
\code{\linkS4class{FuturesCFTC}} objects are returned by the
\code{\link[pullit]{BBG_futures_CFTC}} function in the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{update_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the rebalancing frequency. Must be one of 'year', 'semester',
'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{ranking_period}{a scalar \code{\link[base]{integer}} vector.
Specifies number of periods in term of \code{update_frequency}
looking backward for average CHP calculation. Defaults to 1 where
sort is done on last observation only.}

\item{long_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for long positions. Default: 0.5.}

\item{weighted}{a scalar \code{\link[base]{logical}} vector.
If 'TRUE' adjusts portoflio weights with respect to pressure, else
equal weights are used. Defaults to 'FALSE'.}
}
\value{
An S4 object of class \code{\linkS4class{CHPFactor}}.
}
\description{
Provided with futures contract front price and term
  structure aggregated position data from Bloomberg retrieved with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}},
  construct commercial hedging pressure factor. The futures commercial
  hedging pressure factor is based on the well-known hedging pressure-based
  theory \insertCite{anderson_hedger_1983,chang_returns_1985,cootner_returns_1960,dusak_futures_1973,hicks_value_1939,hirshleifer_risk_1988,hirshleifer_determinants_1989,hirshleifer_hedging_1990,kolb_is_1992,keynes_treatise_1930}{factorem}
  which postulates that futures prices for a given commodity are inversely
  related to the extent that commercial hedgers are short or long and the
  mimicking portfolio here aims at capturing the impact of hedging pressure
  as a systemic factor \insertCite{basu_capturing_2013}{factorem}.
}
\references{
\insertAllCited{}
}
