% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{CHP_factor}
\alias{CHP_factor}
\title{Construct commercial hedging pressure (CHP) factor}
\usage{
CHP_factor(price_data, CHP_data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)
}
\arguments{
\item{price_data}{an S4 object of class \code{\linkS4class{FuturesTS}}. \code{\linkS4class{FuturesTS}} objects are returned by the
\code{\link[pullit]{bbg_futures_TS}} function in the \code{pullit} package.}

\item{CHP_data}{an S4 object of class \code{\linkS4class{FuturesCFTC}}. \code{\linkS4class{FuturesCFTC}} objects are returned by the
\code{\link[pullit]{bbg_futures_CFTC}} function in the \code{pullit} package.}

\item{update_frequency}{a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', 'semester', 'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{ranking_period}{a scalar integer vector. Specifies number of periods in term of \code{update_frequency} looking backward for CHP average calculation.}

\item{long_threshold}{a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{CHPFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector: "CHP factor".}
    \item{\code{returns}: a tibble with columns \code{date}, \code{long} , \code{short} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset supplied for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Sorts inverse CHP.
}
