% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{CHPFactor-class}
\alias{CHPFactor-class}
\title{S4 class for commercial hedging pressure (CHP) factor objects}
\description{
S4 class for commercial hedging pressure (CHP) factor objects
}
