% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{methods}
\name{data,AssetPricingFactor-method}
\alias{data,AssetPricingFactor-method}
\title{Accessor method for factor data construction}
\usage{
\S4method{data}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{An S4 object of class 'AssetPricingFactor'.}
}
\value{
A tibble containing the original dataset used for factor construction.
}
\description{
Accessor method for factor data construction
}
