% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{dataset,AssetPricingFactor-method}
\alias{dataset,AssetPricingFactor-method}
\title{Accessor method for factor data construction}
\usage{
\S4method{dataset}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble containing the original dataset used for factor construction.
}
\description{
Accessor method for factor data construction
}
