% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset}
\alias{dataset}
\title{Generic method for accessing factor data construction}
\usage{
dataset(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble containing the original dataset used for factor construction.
}
\description{
Generic method for accessing factor data construction
}
