% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{factor_returns}
\alias{factor_returns}
\title{Get return time series for factor as well as for long and short legs independently.}
\usage{
factor_returns(data, positions, update_frequency = "month",
  return_frequency = "day", price_variable = "PX_LAST", geometric = TRUE)
}
\arguments{
\item{data}{A dataframe/tibble. Columns must include `name`, `date`, `field` and `value`.}

\item{update_frequency}{A scalar character vector. Specifies the rebalancing frequency. Must be one of "year", "semester", "quarter", "month", "week" or "day". Defaults to "month".}

\item{return_frequency}{A scalar character vector. Specifies the frequency of the returns output. Must be one of "year", "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{price_variable}{A scalar character vector. Specifies the name of the variable hosting asset prices. Must be found in the `field` columns of `data`.}

\item{geometric}{A scalar logical vector. If `TRUE` geometric returns are returned, else arithmetic. Default: `TRUE`.}
}
\value{
A tibble with columns `date`, `long` , `short` and `factor`.
}
\description{
Get return time series for factor as well as for long and short legs independently.
}
