% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{famamcbeth}
\alias{famamcbeth}
\title{Fama-McBeth cross-sectional regression}
\usage{
famamcbeth(assets_returns, factor_returns)
}
\arguments{
\item{assets_returns}{a dataframe with dates in the first column and assets returns
in subsequent columns.}

\item{factor_returns}{a dataframe with dates in the first column and factor returns
in the second.}
}
\value{
An S4 object of class \code{\linkS4class{FamaMcBeth}}.
}
\description{
Given two dataframes of assets returns and factor returns respectively,
  performs a Fama-McBeth two-pass cross-sectional regression.
}
