% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{get_betas}
\alias{get_betas}
\alias{get_betas,FamaMcBeth-method}
\alias{get_betas,FamaMcBeth}
\title{Accesses the \code{betas} slot of S4 objects of class \linkS4class{FamaMcBeth}
  (\href{https://bautheac.github.io/factorem/}{\pkg{factorem}}).}
\usage{
get_betas(object)

\S4method{get_betas}{FamaMcBeth}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FamaMcBeth}.}
}
\value{
A data.table with asset betas.
}
\description{
Access method for the \code{betas} slot of class \linkS4class{FamaMcBeth}
  (\href{https://bautheac.github.io/factorem/}{\pkg{factorem}}).
}
