% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{get_call,AssetPricingFactor-method}
\alias{get_call,AssetPricingFactor-method}
\title{Accessor method for call to factor constructor function}
\usage{
\S4method{get_call}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A scalar character vector showing the original call to the factor constructor function.
}
\description{
Accessor method for call to factor constructor function
}
