% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_call}
\alias{get_call}
\title{Generic method for accessing original call to constructor function}
\usage{
get_call(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A scalar character vector showing the original call to the factor constructor function.
}
\description{
Generic method for accessing original call to constructor function
}
