% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{get_data_original,AssetPricingFactor-method}
\alias{get_data_original,AssetPricingFactor-method}
\title{Accessor method for factor data construction}
\usage{
\S4method{get_data_original}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble containing the original dataset used for factor construction.
}
\description{
Accessor method for factor data construction
}
