% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_data_original}
\alias{get_data_original}
\title{Generic method for accessing factor data construction}
\usage{
get_data_original(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble containing the original dataset used for factor construction.
}
\description{
Generic method for accessing factor data construction
}
