% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{get_lambda}
\alias{get_lambda}
\alias{get_lambda,FamaMcBeth-method}
\alias{get_lambda,FamaMcBeth}
\title{Accesses the \code{lambda} slot of S4 objects of class \linkS4class{FamaMcBeth}
  (\href{https://bautheac.github.io/factorem/}{\pkg{factorem}}).}
\usage{
get_lambda(object)

\S4method{get_lambda}{FamaMcBeth}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FamaMcBeth}.}
}
\value{
An object of class 'lm' with Fama-McBeth
  cross-sectional regression results.
}
\description{
Access method for the \code{lambda} slot of S4 objects of class
  \linkS4class{FamaMcBeth} from the
  \href{https://github.com/bautheac/factorem/}{\pkg{factorem}} package.
}
