% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{get_lambdas}
\alias{get_lambdas}
\alias{get_lambdas,FamaMcBeth-method}
\alias{get_lambdas,FamaMcBeth}
\title{Accesses the \code{lambdas} slot of S4 objects of class \linkS4class{FamaMcBeth}
  (\href{https://bautheac.github.io/factorem/}{\pkg{factorem}}).}
\usage{
get_lambdas(object)

\S4method{get_lambdas}{FamaMcBeth}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FamaMcBeth}.}
}
\value{
A \link[data.table]{data.table} showing Fama-McBeth cross-sectional
regression summary including lambda estimates and corresponding standard errors,
t-stats and p-values.
}
\description{
Access method for the \code{lambdas} slot of S4 objects of class
  \linkS4class{FamaMcBeth} from the
  \href{https://github.com/bautheac/factorem/}{\pkg{factorem}} package.
}
