% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{get_means}
\alias{get_means}
\alias{get_means,FamaMcBeth-method}
\alias{get_means,FamaMcBeth}
\title{Accesses the \code{means} slot of S4 objects of class \linkS4class{FamaMcBeth}
  (\href{https://bautheac.github.io/factorem/}{\pkg{factorem}}).}
\usage{
get_means(object)

\S4method{get_means}{FamaMcBeth}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{FamaMcBeth}.}
}
\value{
A data.table of asset means.
}
\description{
Access method for the \code{means} slot of S4 objects of class
  \linkS4class{FamaMcBeth} from the
  \href{https://github.com/bautheac/factorem/}{\pkg{factorem}} package.
}
