% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{get_name,AssetPricingFactor-method}
\alias{get_name,AssetPricingFactor-method}
\title{Accessor method for factor name}
\usage{
\S4method{get_name}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A unit character vector containing the name of the factor.
}
\description{
Accessor method for factor name
}
