% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_name}
\alias{get_name}
\title{Generic method for accessing factor name}
\usage{
get_name(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A unit character vector containing the name of the factor.
}
\description{
Generic method for accessing factor name
}
