% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{get_parameters,AssetPricingFactor-method}
\alias{get_parameters,AssetPricingFactor-method}
\title{Accessor method for factor parameters}
\usage{
\S4method{get_parameters}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble containing the original parameters supplied for factor construction.
}
\description{
Accessor method for factor parameters
}
