% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{get_positions,AssetPricingFactor-method}
\alias{get_positions,AssetPricingFactor-method}
\title{Accessor method for factor positions}
\usage{
\S4method{get_positions}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor positions.
}
\description{
Accessor method for factor positions
}
