% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_positions}
\alias{get_positions}
\title{Generic method for accessing positions}
\usage{
get_positions(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor positions.
}
\description{
Generic method for accessing positions
}
