% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{get_returns,AssetPricingFactor-method}
\alias{get_returns,AssetPricingFactor-method}
\title{Accessor method for factor returns}
\usage{
\S4method{get_returns}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor returns.
}
\description{
Accessor method for factor returns
}
