% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_returns}
\alias{get_returns}
\title{Generic method for accessing factor returns}
\usage{
get_returns(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor returns.
}
\description{
Generic method for accessing factor returns
}
