% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{market_factor}
\alias{market_factor}
\alias{market_factor,FuturesTS-method}
\alias{market_factor,AssetPricingFactor}
\alias{market_factor,EquityMarket-method}
\title{Market factor}
\usage{
market_factor(data, return_frequency, long)

\S4method{market_factor}{FuturesTS}(data, return_frequency = "month",
  long = TRUE)

\S4method{market_factor}{EquityMarket}(data, return_frequency = "month",
  long = T)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}} or
\code{\linkS4class{EquityMarket}}. \code{\linkS4class{FuturesTS}} and
\code{\linkS4class{EquityMarket}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} and
\code{\link[pullit]{BBG_equity_market}} functions from the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package
respectively.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{long}{a scalar logical vector. If \code{TRUE} long only, else
short only. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}}.
}
\description{
Provided with futures contract front or equity price data
  from Bloomberg retrieved with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}, construct
  market factor.
}
