% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{market_factor}
\alias{market_factor}
\title{Construct market factor}
\usage{
market_factor(data, return_frequency = "day", long = TRUE,
  geometric = TRUE)
}
\arguments{
\item{data}{A dataframe/tibble. Columns must include `active_contract_ticker`, `date`, `TS_position`, `field` and `value`.}

\item{return_frequency}{A scalar character vector. Specifies the frequency of the returns output. Must be one of "year", "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{long}{A scalar logical vector. If `TRUE` long only, else short only. Default: `TRUE`.}

\item{geometric}{A scalar logical vector. If `TRUE` geometric returns are returned, else arithmetic. Default: `TRUE`.}
}
\value{
An S4 object of class \linkS4class{AssetPricingFactor} with slots:
  \itemize{
    \item{name: a scalar character vector: "market factor".}
    \item{returns: a tibble with columns `date` and `factor`.}
    \item{positions: a tibble with columns `date`, `name` and `position`.}
    \item{data: a tibble containing the original dataset used for factor construction.}
    \item{params: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Given a reference dataset and a set of parameters.
}
