% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{momentum_factor}
\alias{momentum_factor}
\alias{momentum_factor,FuturesTS-method}
\alias{momentum_factor,EquityMarket-method}
\title{Construct momentum factor}
\usage{
momentum_factor(data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)

\S4method{momentum_factor}{FuturesTS}(data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)

\S4method{momentum_factor}{EquityMarket}(data,
  update_frequency = "month", return_frequency = "day",
  ranking_period = 0L, long_threshold = 0.5, short_threshold = 0.5,
  geometric = TRUE)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}} or \code{\linkS4class{EquityMarket}}. \code{\linkS4class{FuturesTS}} and
\code{\linkS4class{EquityMarket}} objects are returned by the \code{\link[pullit]{bbg_futures_TS}} and \code{\link[pullit]{bbg_equity_market}}
functions from the \code{pullit} package respectively.}

\item{update_frequency}{a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', "semester", "quarter", "month" or "week". Defaults to "month".}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{ranking_period}{a scalar integer vector. Specifies number of periods in term of \code{update_frequency} looking backward for averaging \code{sort_variable}.}

\item{long_threshold}{a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{MomentumFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector showing the factor name.}
    \item{\code{returns}: a tibble with columns \code{date}, \code{long}, \code{short} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset supplied for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Sorts on past returns.
}
\section{Methods (by class)}{
\itemize{
\item \code{FuturesTS}: futures momentum factor.

\item \code{EquityMarket}: equity momentum factor.
}}

