% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{methods}
\name{name,AssetPricingFactor-method}
\alias{name,AssetPricingFactor-method}
\title{Accessor method for factor name}
\usage{
\S4method{name}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{An S4 object of class 'AssetPricingFactor'.}
}
\value{
A unit character vector containing the name of the factor.
}
\description{
Accessor method for factor name
}
