% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{OI_aggregate_factor}
\alias{OI_aggregate_factor}
\alias{OI_aggregate_factor,FuturesTS,FuturesAggregate-method}
\alias{OI_aggregate_factor,AssetPricingFactor}
\title{Construct futures open interest growth (OI) aggregate factor}
\usage{
OI_aggregate_factor(price_data, aggregate_data, update_frequency,
  return_frequency, ranking_period, long_threshold, short_threshold,
  geometric)

\S4method{OI_aggregate_factor}{FuturesTS,FuturesAggregate}(price_data,
  aggregate_data, update_frequency = "month", return_frequency = "day",
  ranking_period = 0L, long_threshold = 0.5, short_threshold = 0.5,
  geometric = TRUE)
}
\arguments{
\item{price_data}{an S4 object of class \code{\linkS4class{FuturesTS}}.
\code{\linkS4class{FuturesTS}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} function from the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{aggregate_data}{an S4 object of class \code{\linkS4class{FuturesAggregate}}.
\code{\linkS4class{FuturesAggregate}} objects are returned by the
\code{\link[pullit]{BBG_futures_aggregate}} function from the
\code{pullit} package.}

\item{update_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the rebalancing frequency. Must be one of 'year', 'semester',
'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{ranking_period}{a scalar \code{\link[base]{integer}} vector.
Specifies number of periods in term of \code{update_frequency}
looking backward for aggregate open interest average growth calculation.
Defaults to 0 where sort is done on last observation only.}

\item{long_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{a scalar \code{\link[base]{logical}} vector. If \code{TRUE}
geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{OIFactor}}.
}
\description{
Provided with futures contract front price and aggregate open
  interest data from Bloomberg retrieved with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}, construct
  aggregate open interest growth (OI) factor.

Basu & Bautheac.
}
