% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{OI_factor}
\alias{OI_factor}
\alias{OI_factor,FuturesTS-method}
\alias{OI_factor,FuturesAggregate-method}
\title{Construct futures open interest growth (OI) factor}
\usage{
OI_factor(data, update_frequency, return_frequency, ranking_period,
  long_threshold, short_threshold, geometric)

\S4method{OI_factor}{FuturesTS}(data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)

\S4method{OI_factor}{FuturesAggregate}(data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}} or \code{\linkS4class{FuturesAggregate}}. \code{\linkS4class{FuturesTS}} and
\code{\linkS4class{FuturesAggregate}} objects are returned by the \code{\link[pullit]{bbg_futures_TS}} and \code{\link[pullit]{bbg_futures_aggregate}}
functions from the \code{pullit} package respectively.}

\item{update_frequency}{a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', "semester", "quarter", "month" or "week". Defaults to "month".}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{ranking_period}{a scalar integer vector. Specifies number of periods in term of \code{update_frequency} looking backward for averaging \code{sort_variable}.}

\item{long_threshold}{a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{OIFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector showing the factor name.}
    \item{\code{returns}: a tibble with columns \code{date}, \code{long}, \code{short} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset supplied for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Following Hong & Yogo (nearby) as well as Basu & Bautheac (aggregate).
}
\section{Methods (by class)}{
\itemize{
\item \code{FuturesTS}: nearby open interest growth factor.

\item \code{FuturesAggregate}: aggregate open interest growth factor.
}}

