% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{OI_nearby_factor}
\alias{OI_nearby_factor}
\alias{OI_nearby_factor,FuturesTS-method}
\alias{OI_nearby_factor,AssetPricingFactor}
\title{Open interest growth (OI) nearby factor}
\usage{
OI_nearby_factor(data, update_frequency, return_frequency, ranking_period,
  long_threshold, short_threshold, geometric)

\S4method{OI_nearby_factor}{FuturesTS}(data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}}.
\code{\linkS4class{FuturesTS}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} function from the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{update_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the rebalancing frequency. Must be one of 'year', 'semester',
'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{ranking_period}{a scalar \code{\link[base]{integer}} vector.
Specifies number of periods in term of \code{update_frequency}
looking backward for nearby open interest average growth calculation.
Defaults to 0 where sort is done on last observation only.}

\item{long_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{a scalar \code{\link[base]{logical}} vector. If \code{TRUE}
geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{OIFactor}}.
}
\description{
Provided with futures contract front price and open interest data from
  Bloomberg retrieved with \href{https://github.com/bautheac/pullit/}{\pkg{pullit}},
  construct open interest growth (OI) nearby factor.

Following Hong & Yogo.
}
