% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{methods}
\name{parameters,AssetPricingFactor-method}
\alias{parameters,AssetPricingFactor-method}
\title{Accessor method for factor parameters}
\usage{
\S4method{parameters}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{An S4 object of class 'AssetPricingFactor'.}
}
\value{
A tibble containing the original parameters supplied for factor construction.
}
\description{
Accessor method for factor parameters
}
