% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\name{plot_performance}
\alias{plot_performance}
\title{Generic method factor performance summary plot by leg}
\usage{
plot_performance(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A line plot showing the wealth index for the factor itself as well as that of both long and short legs separately.
}
\description{
Generic method factor performance summary plot by leg
}
