% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_positions}
\alias{plot_positions}
\title{Generic method factor positions summary plot by leg}
\usage{
plot_positions(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A bar plot showing the proportion of time individual names appear in the factor.
}
\description{
Generic method factor positions summary plot by leg
}
