% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\name{positions}
\alias{positions}
\title{Generic method for accessing positions}
\usage{
positions(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor positions.
}
\description{
Generic method for accessing positions
}
