% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{pressure_factor}
\alias{pressure_factor}
\alias{pressure_factor,FuturesTS,FuturesCFTC-method}
\alias{pressure_factor,AssetPricingFactor}
\title{Market participant pressure factor}
\usage{
pressure_factor(price_data, position_data, format = "disaggregated",
  underlying = "futures only", unit = "contracts",
  participant = "producer/merchant/processor/user",
  update_frequency = "month", return_frequency = "day",
  ranking_period = 6L, long_threshold = 0.5, short_threshold = 0.5,
  weighted = F)

\S4method{pressure_factor}{FuturesTS,FuturesCFTC}(price_data,
  position_data, format = "disaggregated", underlying = "futures only",
  unit = "contracts", participant = "producer/merchant/processor/user",
  update_frequency = "month", return_frequency = "day",
  ranking_period = 6L, long_threshold = 0.5, short_threshold = 0.5,
  weighted = F)
}
\arguments{
\item{price_data}{an S4 object of class \code{\linkS4class{FuturesTS}}.
\code{\linkS4class{FuturesTS}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} function in the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{position_data}{an S4 object of class \code{\linkS4class{FuturesCFTC}}.
\code{\linkS4class{FuturesCFTC}} objects are returned by the
\code{\link[pullit]{BBG_futures_CFTC}} function in the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{format}{a scalar \code{\link[base]{character}} vector.
Specifies the CFTC report format to use for calculating participant
pressure: 'disaggregated', 'legacy', 'supplemental' or
'traders in financial futures'. Defaults to 'disaggregated'.}

\item{underlying}{a scalar \code{\link[base]{character}} vector.
Specifies the underlying instrument type to use for calculating participant
pressure: 'futures only' or 'futures & options'. Defaults to 'futures only'.}

\item{unit}{a scalar \code{\link[base]{character}} vector.
Specifies the unit to use for calculating participant
pressure: 'contracts' or 'traders'. Defaults to 'contracts'.}

\item{participant}{a scalar \code{\link[base]{character}} vector.
Specifies the participant to use for calculating market pressure.
Format specific: 'producer/merchant/processor/user', 'managed money',
'swap dealers' or 'other reportables' (disaggregated); 'commercial',
'non-commercial', 'non-reportable' or 'total' (legacy);
'commercial - non-CIT', 'non-commercial - non-CIT', or 'index traders'
(supplemental); 'asset manager/institutional', 'dealer/intermediary',
'leveraged funds', 'other reportables' (traders in financial futures).}

\item{update_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the rebalancing frequency. Must be one of 'year', 'semester',
'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{ranking_period}{a scalar \code{\link[base]{integer}} vector.
Specifies number of periods in term of \code{update_frequency}
looking backward for average CHP calculation. Defaults to 1 where
sort is done on last observation only.}

\item{long_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar \code{\link[base]{numeric}} vector.
Specifies the threshold for long positions. Default: 0.5.}

\item{weighted}{a scalar \code{\link[base]{logical}} vector.
If 'TRUE' adjusts portoflio weights with respect to pressure, else
equal weights are used. Defaults to 'FALSE'.}
}
\value{
An S4 object of class \code{\linkS4class{PressureFactor}}.
}
\description{
Provided with futures contract front price and term
  structure aggregated position data from Bloomberg retrieved with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}},
  constructs market participant pressure factors.
}
