% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\name{returns}
\alias{returns}
\title{Generic method for accessing factor returns}
\usage{
returns(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor returns.
}
\description{
Generic method for accessing factor returns
}
