% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{summary,AssetPricingFactor-method}
\alias{summary,AssetPricingFactor-method}
\title{Summary of factor returns by month and year}
\usage{
\S4method{summary}{AssetPricingFactor}(object, leg = "factor")
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}

\item{leg}{A character vector. 'long', 'short' and 'factor' display the returns for the long, short and the factor itself respectively.}
}
\value{
A dataframe with months as colums and years as rows. Column total shows the cumulated return for the corresponding year.
}
\description{
Summary of factor returns by month and year
}
