% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{TS_factor}
\alias{TS_factor}
\alias{TS_factor,FuturesTS-method}
\alias{TS_factor,AssetPricingFactor}
\title{Term structure factor}
\usage{
TS_factor(
  data,
  update_frequency = "month",
  return_frequency = "day",
  front = 1L,
  back = 2L,
  ranking_period = 1L,
  long_threshold = 0.5,
  short_threshold = 0.5,
  weighted = F
)

\S4method{TS_factor}{FuturesTS}(
  data,
  update_frequency = "month",
  return_frequency = "day",
  front = 1L,
  back = 2L,
  ranking_period = 1L,
  long_threshold = 0.5,
  short_threshold = 0.5,
  weighted = F
)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}}.
\code{\linkS4class{FuturesTS}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} function from the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.}

\item{update_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the rebalancing frequency. Must be one of 'year', 'semester',
'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{front}{a scalar \code{\link[base]{integer}} vector. Specifies the
term structure position to use as front contract in roll yield
(sort variable) calculation. Defaults to 1.}

\item{back}{a scalar \code{\link[base]{integer}} vector. Specifies the
term structure position to use as back contract in roll yield
(sort variable) calculation. Defaults to 2.}

\item{ranking_period}{a scalar \code{\link[base]{integer}} vector. Specifies
number of periods in term of \code{update_frequency} looking backward for
average roll yield calculation. Defaults to 1 where sort is done on last
observation only.}

\item{long_threshold}{a scalar \code{\link[base]{numeric}} vector. Specifies
the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar \code{\link[base]{numeric}} vector. Specifies
the threshold for long positions. Default: 0.5.}

\item{weighted}{a scalar \code{\link[base]{logical}} vector.
If 'TRUE' adjusts portoflio weights with respect to sorting variable
values, else equal weights are used. Defaults to 'FALSE'.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}}.
}
\description{
Provided with futures contract front price data
  from Bloomberg retrieved with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}, construct
  term structure factor. The futures term structure factor
  \insertCite{szymanowska_anatomy_2014,fuertes_commodity_2015}{factorem}
  is concerned with the shape (steepness) of the futures term structure
  and sorts on roll yield.
}
\references{
\insertAllCited{}
}
