% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{TS_factor}
\alias{TS_factor}
\alias{TS_factor,FuturesTS-method}
\title{Construct term structure factor}
\usage{
TS_factor(data, update_frequency, return_frequency, front, back,
  ranking_period, long_threshold, short_threshold, geometric)

\S4method{TS_factor}{FuturesTS}(data, update_frequency = "month",
  return_frequency = "day", ranking_period = 0L,
  long_threshold = 0.5, short_threshold = 0.5, geometric = TRUE)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}}. \code{\linkS4class{FuturesTS}} objects are returned by the \code{\link[pullit]{bbg_futures_TS}}
function from the \code{pullit} package.}

\item{update_frequency}{a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', "semester", "quarter", "month" or "week". Defaults to "month".}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{front}{a scalar integer vector. Specifies the term structure position to use as front contract in roll yield (sort variable) calculation. Defaults to 1.}

\item{back}{a scalar integer vector. Specifies the term structure position to use as back contract in roll yield (sort variable) calculation. Defaults to 2.}

\item{ranking_period}{a scalar integer vector. Specifies number of periods in term of \code{update_frequency} looking backward for averaging \code{sort_variable}.}

\item{long_threshold}{a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector showing the factor name.}
    \item{\code{returns}: a tibble with columns \code{date}, \code{long} , \code{short} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset used for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Sorts on relative difference between specified front and back.
}
\section{Methods (by class)}{
\itemize{
\item \code{FuturesTS}: futures term structure factor.
}}

