% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{class}
\name{AssetPricingFactor-class}
\alias{AssetPricingFactor-class}
\alias{.AssetPricingFactor}
\title{S4 class for asset pricing factor objects}
\description{
S4 class for asset pricing factor objects
}
