% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{CHP_factor}
\alias{CHP_factor}
\title{Construct CHP factor}
\usage{
CHP_factor(price_data, CHP_data, update_frequency = c("year", "semester",
  "quarter", "month", "week"), return_frequency = c("year", "semester",
  "quarter", "month", "week", "day"), sort_levels = FALSE,
  ranking_period = 0L, long_threshold = 0.5, short_threshold = 0.5,
  geometric = TRUE)
}
\arguments{
\item{price_data}{A dataframe/tibble. Columns must include `name`, `date`, `field` and `value`. `field` must contain `PX_LAST` variable.}

\item{CHP_data}{A dataframe/tibble. Columns must include:
\itemize{
  \item{format: must contain `legacy`.}
  \item{underlying: must contain `futures only`.}
  \item{unit: must contain `# positions`.}
  \item{participant: must contain `commercial`.}
  \item{position: must contain `long` & `short`.}
  \item{PX_LAST: contains the corresponding values.}
}}

\item{update_frequency}{A unit length character vector. Specifies the rebalancing frequency. Must be one of "year", "semester", "quarter", "month" or "week".}

\item{return_frequency}{A unit length character vector. Specifies the frequency of the returns output. Must be one of "year", "semester", "quarter", "month", "week" or "day".}

\item{sort_levels}{Logical. If `TRUE`, sort is done on `sort_variable`'s levels, else on relative changes. Default: `FALSE`.}

\item{ranking_period}{A unit length integer vector. Specifies number of periods in term of `update_frequency` looking backward for averaging `sort_variable`.}

\item{long_threshold}{A unit length numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{A unit length numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{geometric}{Logical. If `TRUE` geometric returns are returned, else arithmetic. Default: `TRUE`.}
}
\value{
an \linkS4class{AssetPricingFactor} object with slots:
  \itemize{
    \item{name: a unit character vector: "CHP factor".}
    \item{returns: a tibble with columns `date`, `long` , `short` and `factor`.}
    \item{positions: a tibble with columns `date`, `name` and `position`.}
    \item{data: a tibble containing the original dataset used for factor construction.}
    \item{params: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Construct CHP factor
}
