% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{factor_positions}
\alias{factor_positions}
\title{Get time series of long and short factor positions}
\usage{
factor_positions(data, update_frequency = c("year", "semester", "quarter",
  "month", "week", "day"), sort_variable = "PX_LAST", sort_levels = FALSE,
  ranking_period = 0L, long_threshold = 0.5, short_threshold = 0.5)
}
\arguments{
\item{data}{A dataframe/tibble. Columns must include `name`, `date`, `field` and `value`.}

\item{update_frequency}{A unit length character vector. Specifies the rebalancing frequency. Must be one of "year", "semester", "quarter", "month", "week" or "day".}

\item{sort_variable}{A unit length character vector. Specifies the name of the variable to use for sorting. Must be found in the `field` columns of `data`.}

\item{sort_levels}{Logical. If `TRUE`, sort is done on `sort_variable`'s levels, else on relative changes. Default: `FALSE`.}

\item{ranking_period}{A unit length integer vector. Specifies number of periods in term of `update_frequency` looking backward for averaging `sort_variable`}

\item{long_threshold}{A unit length numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{short_threshold}{A unit length numeric vector. Specifies the threshold for short positions. Default: 0.5.}
}
\value{
A tibble with columns `date`, `name` and `position`.
}
\description{
Get time series of long and short factor positions
}
