% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{factorem}
\alias{factorem}
\title{Construct an asset pricing factor}
\usage{
factorem(
  name = "",
  data,
  update_frequency = "month",
  return_frequency = "day",
  price_variable = "PX_LAST",
  sort_variable = "PX_LAST",
  sort_levels = T,
  weighted = T,
  ranking_period = 1L,
  long_threshold = 0.5,
  short_threshold = 0.5
)
}
\arguments{
\item{name}{a scalar character vector specifying the name to use for the factor.}

\item{data}{a dataframe/tibble. Columns must include \code{ticker}, \code{date},
\code{field} and \code{value}.}

\item{update_frequency}{a scalar character vector. Specifies the rebalancing
frequency. Must be one of 'year', 'semester', 'quarter', 'month', 'week' or
'day'. Defaults to 'month'.}

\item{return_frequency}{a scalar character vector. Specifies the frequency of
the returns output. Must be one of 'year', 'semester', 'quarter', 'month',
'week' or 'day'. Defaults to 'day'.}

\item{price_variable}{a scalar character vector. Specifies the name of the variable
hosting asset prices. Must be found in the \code{field} columns of \code{data}.}

\item{sort_variable}{a scalar character vector. Specifies the name of the variable
to use for sorting. Must be found in the \code{field} columns of \code{data}.}

\item{sort_levels}{a scalar logical vector. If \code{TRUE}, sort is done on
\code{sort_variable}'s levels, else on relative changes. Default: \code{TRUE}.}

\item{weighted}{a scalar \code{\link[base]{logical}} vector.
If \code{TRUE} adjusts portoflio weights with respect to sorting variable, else
equal weights are used. Defaults to \code{FALSE}.}

\item{ranking_period}{a scalar integer vector. Specifies number of periods in term
of \code{update_frequency} looking backward for averaging \code{sort_variable}.
Default: 1 (sort on last observation).}

\item{long_threshold}{a scalar numeric vector. Specifies the threshold for short
positions. Default: 0.5.}

\item{short_threshold}{a scalar numeric vector. Specifies the threshold for long
positions. Default: 0.5.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}}.
}
\description{
Given a reference dataset and a set of parameters, construct the desired asset
  pricing factor and returns its return and position time series by leg.
}
