% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data,AssetPricingFactor-method}
\alias{get_data,AssetPricingFactor}
\title{Accesses the \code{data} slot of S4 objects of class \linkS4class{AssetPricingFactor}
  (\href{https://bautheac.github.io/factorem/}{\pkg{factorem}}).}
\usage{
get_data(object)

\S4method{get_data}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A \link[data.table]{data.table} showing for each Bloomberg ticker and
  Bloomberg field the data contained in the \code{object} provided.
}
\description{
Access method for the \code{data} slot of S4 objects of class
  \linkS4class{AssetPricingFactor} from the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
