% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{market_factor}
\alias{market_factor}
\alias{market_factor,FuturesTS-method}
\alias{market_factor,EquityMarket-method}
\title{Construct market factor}
\usage{
market_factor(data, return_frequency = "day", long = TRUE,
  geometric = TRUE)

\S4method{market_factor}{FuturesTS}(data, return_frequency = "day",
  long = TRUE, geometric = TRUE)

\S4method{market_factor}{EquityMarket}(data, return_frequency = "day",
  long = TRUE, geometric = TRUE)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}} or \code{\linkS4class{EquityMarket}}. \code{\linkS4class{FuturesTS}} and
\code{\linkS4class{EquityMarket}} objects are returned by the \code{\link[pullit]{bbg_futures_TS}} and \code{\link[pullit]{bbg_equity_market}}
functions from the \code{pullit} package respectively.}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{long}{a scalar logical vector. If \code{TRUE} long only, else short only. Default: \code{TRUE}.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector: "market factor".}
    \item{\code{returns}: a tibble with columns \code{date} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset used for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Long only or short only equally weighted portfolio.
}
\section{Methods (by class)}{
\itemize{
\item \code{FuturesTS}: futures nearby market factor.

\item \code{EquityMarket}: equity market factor.
}}

