% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{market_factor}
\alias{market_factor}
\title{Construct market factor}
\usage{
market_factor(data, return_frequency = "day", long = TRUE,
  geometric = TRUE)
}
\arguments{
\item{data}{a dataframe/tibble. Columns must include \code{active contract ticker}, \code{date}, \code{TS position}, \code{field} and \code{value}.}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{long}{a scalar logical vector. If \code{TRUE} long only, else short only. Default: \code{TRUE}.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector: "market factor".}
    \item{\code{returns}: a tibble with columns \code{date} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset used for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Given a reference dataset and a set of parameters.
}
