% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.r
\docType{methods}
\name{momentum_factor}
\alias{momentum_factor}
\alias{momentum_factor,FuturesTS-method}
\alias{momentum_factor,AssetPricingFactor}
\alias{momentum_factor,EquityMarket-method}
\title{Construct momentum factor}
\usage{
momentum_factor(data, update_frequency = "week",
  return_frequency = "day", ranking_period = 4L,
  long_threshold = 0.5, short_threshold = 0.5, weighted = F,
  risk_adjusted = F)

\S4method{momentum_factor}{FuturesTS}(data, update_frequency = "week",
  return_frequency = "day", ranking_period = 4L,
  long_threshold = 0.5, short_threshold = 0.5, weighted = F,
  risk_adjusted = F)

\S4method{momentum_factor}{FuturesTS}(data, update_frequency = "week",
  return_frequency = "day", ranking_period = 4L,
  long_threshold = 0.5, short_threshold = 0.5, weighted = F,
  risk_adjusted = F)

\S4method{momentum_factor}{EquityMarket}(data, update_frequency = "week",
  return_frequency = "day", ranking_period = 4L,
  long_threshold = 0.5, short_threshold = 0.5, weighted = F,
  risk_adjusted = F)
}
\arguments{
\item{data}{an S4 object of class \code{\linkS4class{FuturesTS}} or
\code{\linkS4class{EquityMarket}}. \code{\linkS4class{FuturesTS}} and
\code{\linkS4class{EquityMarket}} objects are returned by the
\code{\link[pullit]{BBG_futures_TS}} and \code{\link[pullit]{BBG_equity_market}}
functions from the
\href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package respectively.}

\item{update_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the rebalancing frequency. Must be one of 'year', 'semester',
'quarter', 'month' or 'week'. Defaults to 'month'.}

\item{return_frequency}{a scalar \code{\link[base]{character}} vector.
Specifies the frequency of the returns output. Must be one of 'year',
'semester', 'quarter', 'month', 'week' or 'day'. Defaults to 'day'.}

\item{ranking_period}{a scalar \code{\link[base]{integer}} vector. Specifies
number of periods in term of \code{update_frequency} looking backward for
average price return calculation. Defaults to 1 where sort is done on last
observed return.}

\item{long_threshold}{a scalar \code{\link[base]{numeric}} vector. Specifies
the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar \code{\link[base]{numeric}} vector. Specifies
the threshold for long positions. Default: 0.5.}

\item{weighted}{a scalar \code{\link[base]{logical}} vector.
If 'TRUE' adjusts portoflio weights with respect to sorting variable
values, else equal weights are used. Defaults to 'FALSE'.}

\item{risk_adjusted}{a scalar \code{\link[base]{logical}} vector.
If 'TRUE' returns a reward/risk mometum factor, else returns a classic
momentum factor. Defaults to 'FALSE'.}
}
\value{
An S4 object of class \code{\linkS4class{MomentumFactor}}.
}
\description{
Provided with futures contract front or equity price data
  from Bloomberg retrieved with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}, construct
  momentum factor.

Sorts on past returns.
}
