% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{OI_factor}
\alias{OI_factor}
\title{Construct OI factor}
\usage{
OI_factor(data, update_frequency = "month", return_frequency = "day",
  ranking_period = 0L, long_threshold = 0.5, short_threshold = 0.5,
  aggregate = FALSE, geometric = TRUE)
}
\arguments{
\item{data}{a dataframe/tibble. Columns must include \code{active contract ticker}, \code{date}, \code{TS position}, \code{field} and \code{value}. \code{field} must contain \code{PX_LAST} and \code{OPEN_INT} variables.}

\item{update_frequency}{a scalar character vector. Specifies the rebalancing frequency. Must be one of 'year', "semester", "quarter", "month" or "week". Defaults to "month".}

\item{return_frequency}{a scalar character vector. Specifies the frequency of the returns output. Must be one of 'year', "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{ranking_period}{a scalar integer vector. Specifies number of periods in term of \code{update_frequency} looking backward for averaging \code{sort_variable}.}

\item{long_threshold}{a scalar numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{a scalar numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{aggregate}{a scalar logical vector. If \code{TRUE} open interest is aggregated over the whole term structure supplied in \code{data}. Defaults to \code{FALSE}.}

\item{geometric}{a scalar logical vector. If \code{TRUE} geometric returns are returned, else arithmetic. Default: \code{TRUE}.}
}
\value{
An S4 object of class \code{\linkS4class{AssetPricingFactor}} with slots:
  \itemize{
    \item{\code{name}: a scalar character vector: "OI factor".}
    \item{\code{returns}: a tibble with columns \code{date}, \code{long}, \code{short} and \code{factor}.}
    \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
    \item{\code{data}: a tibble containing the original dataset used for factor construction.}
    \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Given a reference dataset and a set of parameters.
}
