% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{OI_factor}
\alias{OI_factor}
\title{Construct OI factor}
\usage{
OI_factor(data, update_frequency = "month", return_frequency = "day",
  ranking_period = 0L, long_threshold = 0.5, short_threshold = 0.5,
  aggregate = FALSE, geometric = TRUE)
}
\arguments{
\item{data}{A dataframe/tibble. Columns must include `active_contract_ticker`, `date`, `field` and `value`. `field` must contain `PX_LAST` and `OPEN_INT` variables.}

\item{update_frequency}{A unit length character vector. Specifies the rebalancing frequency. Must be one of "year", "semester", "quarter", "month" or "week". Defaults to "month".}

\item{return_frequency}{A unit length character vector. Specifies the frequency of the returns output. Must be one of "year", "semester", "quarter", "month", "week" or "day". Defaults to "day".}

\item{ranking_period}{A unit length integer vector. Specifies number of periods in term of `update_frequency` looking backward for averaging `sort_variable`.}

\item{long_threshold}{A unit length numeric vector. Specifies the threshold for short positions. Default: 0.5.}

\item{short_threshold}{A unit length numeric vector. Specifies the threshold for long positions. Default: 0.5.}

\item{aggregate}{Logical. If `TRUE` open interest is aggregated over the whole term structure supplied in `data`. Defaults to `FALSE`.}

\item{geometric}{Logical. If `TRUE` geometric returns are returned, else arithmetic. Default: `TRUE`.}
}
\value{
An \linkS4class{AssetPricingFactor} object with slots:
  \itemize{
    \item{name: a unit character vector: "OI factor".}
    \item{returns: a tibble with columns `date`, `long` , `short` and `factor`.}
    \item{positions: a tibble with columns `date`, `name` and `position`.}
    \item{data: a tibble containing the original dataset used for factor construction.}
    \item{params: a tibble containing the original parameters supplied for factor construction.}
  }
}
\description{
Given a reference dataset and a set of parameters.
}
