% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{methods}
\name{plot_performance,AssetPricingFactor-method}
\alias{plot_performance,AssetPricingFactor-method}
\title{Summary plot of factor performance by leg}
\usage{
\S4method{plot_performance}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{An S4 object of class 'AssetPricingFactor'.}
}
\value{
A line plot showing the wealth index for the factor itself as well as that of both long and short legs separately.
}
\description{
Summary plot of factor performance by leg
}
