% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{plot_positions,AssetPricingFactor-method}
\alias{plot_positions,AssetPricingFactor-method}
\title{Summary plot of factor positions by leg}
\usage{
\S4method{plot_positions}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A bar plot showing the proportion of time individual names appear in the factor.
}
\description{
Summary plot of factor positions by leg
}
