% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{positions,AssetPricingFactor-method}
\alias{positions,AssetPricingFactor-method}
\title{Accessor method for factor positions}
\usage{
\S4method{positions}{AssetPricingFactor}(factor)
}
\arguments{
\item{factor}{an S4 object of class \linkS4class{AssetPricingFactor}.}
}
\value{
A tibble of factor positions.
}
\description{
Accessor method for factor positions
}
