% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.r
\docType{methods}
\name{show,AssetPricingFactor-method}
\alias{show,AssetPricingFactor-method}
\title{Show method for S4 object of class `AssetPricingFactor`.}
\usage{
\S4method{show}{AssetPricingFactor}(object)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{AssetPricingFactor} with slots:
\itemize{
  \item{\code{name}: a scalar character vector specifying the name to use for the factor.}
  \item{\code{returns}: a tibble with columns \code{date}, \code{long} , \code{short} and \code{factor}.}
  \item{\code{positions}: a tibble with columns \code{date}, \code{name} and \code{position}.}
  \item{\code{data}: a tibble containing the original dataset used for factor construction.}
  \item{\code{params}: a tibble containing the original parameters supplied for factor construction.}
}}
}
\description{
Show method for S4 object of class `AssetPricingFactor`.
}
