% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{show,AssetPricingFactor-method}
\alias{show,AssetPricingFactor-method}
\alias{show,AssetPricingFactor}
\title{Show method for \href{https://bautheac.github.io/factorem/}{\pkg{factorem}} S4 objects.}
\usage{
\S4method{show}{AssetPricingFactor}(object)
}
\description{
Show method for \href{https://bautheac.github.io/factorem/}{\pkg{factorem}} S4 objects.
}
